Renamed to o3. #35
Added the examples to the spec directory. Last time I checked it was not possible to use jasmine without a spec dir and I did not want nested directories like example/spec. This is simple refactoring.

I made it clear in the readme file that setting the environment variable is only required for testing.  #34