Managed to move the newConstructor to the Class with being non-DRY. I made the Class abstract, so one copy was removed. The second and third copies are in UserError, I will deal with them later.
I added clone and mixin to the Class, and extend is now a combination of these. I removed the extend and mixin functions from the utils, and removed the Class cloning from shallowClone as well. These are Class specific functions, they don't belong to the utils.

Still in need to cut the circular dependency between UserError subclasses and the util functions, but at least the Class creation is removed from the utils. One step closer to separate files. #45