Moved the stack string generation code the a separate Stack class. Moved the frame generation code to this class, so only this class will call the adapter directly, the UserError classes can interact only with this class.
Added platform check, so the code can select adapter depending on the environment. #69 I designed the adapters to bypass the #77 bug/feature, because I don't want to wait for a browserify fix.

Added phantomjs support. #62 It works more or less. #79 #80 I have to work on it a little bit more, but it is better to move on and add Firefox support, since that is my default browser. I think Opera can wait, and probably I should add some feature to the mock, something like using the original frame string if they are in an unknown format. Probably adding some Frame subclass could help. #81